primitive testing
Verifies the html output of sources in test/ to files in expected/.
It isn't the easiest format to manually check, but the html output
is almost purely semantic markup, so this is a good verification of
the extracted model.  This will be somewhat fragile until
better identifiers for links are implemented.